lint passes, transaction types consolidated
